Simplify sd command and fix visor service detection
- Remove network subcommand, make 'sd' the direct command

- Change from 'skywire cli sd network' to 'skywire cli sd'

- Fix visor address parsing to strip port numbers

- Now visor service type shows correctly with transport counts

Before: visor entries showed 0 transports (address mismatch)

After: visor entries show correct transport counts
